fix mem alias bug, fix mergeRegions bug, clean up

+++ tensor_graphs_cpp/core/misc.hpp
@@ -325,8 +325,8 @@ inline std::vector<Region> mergeRegions(const std::vector<Region> &regions)
     if (result.empty())
         return result;
 
-    const size_t rank = result.front().region.size();
-    for (size_t dim = rank - 1; dim >= 0; dim--) // TODO: maybe merge top down? could discover different optimizations in InfinityDomination if we merge different ways?
+    const int rank = result.front().region.size();
+    for (int dim = rank - 1; dim >= 0; dim--) // TODO: maybe merge top down? could discover different optimizations in InfinityDomination if we merge different ways?
     {
         std::vector<Region> next = mergeRegionsAlongDim(result, dim);
         if (encodeRegionList(next) != encodeRegionList(result))
